+++ example/lib/enemy/goblin.dart
@@ -150,7 +150,6 @@ class Goblin extends SimpleEnemy {
         ),
       ),
       lightingConfig: LightingConfig(
-        gameComponent: this,
         radius: width,
         blurBorder: width * 0.5,
       ),