+++ lib/widgets/bonfire_tiled_widget.dart
@@ -4,6 +4,7 @@ import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/game_interface/game_interface.dart';
 import 'package:bonfire/joystick/joystick_controller.dart';
 import 'package:bonfire/player/player.dart';
+import 'package:bonfire/tiled/tiled_world_data.dart';
 import 'package:bonfire/tiled/tiled_world_map.dart';
 import 'package:bonfire/util/game_controller.dart';
 import 'package:flutter/material.dart';
@@ -30,7 +31,7 @@ class BonfireTiledWidget extends StatefulWidget {
 
   const BonfireTiledWidget({
     Key key,
-    @required this.map,
+    this.map,
     this.joystick,
     this.player,
     this.interface,
@@ -61,6 +62,7 @@ class _BonfireTiledWidgetState extends State<BonfireTiledWidget>
   @override
   void didUpdateWidget(BonfireTiledWidget oldWidget) {
     if (widget.constructionMode) {
+      if (widget.map == null) return;
       widget.map.build().then((value) {
         _game.map.updateTiles(value.map.tiles);
 
@@ -94,14 +96,18 @@ class _BonfireTiledWidgetState extends State<BonfireTiledWidget>
   }
 
   void _loadGame() async {
-    final tiled = await widget.map.build();
+    TiledWorldData tiled;
+    if (widget.map != null) {
+      tiled = await widget.map.build();
+    }
+
     _game = RPGGame(
       context: context,
       joystickController: widget.joystick,
       player: widget.player,
       interface: widget.interface,
-      map: tiled.map,
-      components: tiled.components,
+      map: tiled?.map,
+      components: tiled?.components ?? [],
       background: widget.background,
       constructionMode: widget.constructionMode,
       showCollisionArea: widget.showCollisionArea,